#10058 NewAdminBundle: Restringed deletion when using batchDelete
Now batchDelete will delete objects that the user is allowed to delete and leave the rest untouched.

+++ src/Pumukit/NewAdminBundle/Controller/MultimediaObjectController.php
@@ -53,7 +53,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
                 $update_session = false;
             }
         }
- 
+
         if($update_session){
             $this->get('session')->remove('admin/mms/id');
         }
@@ -263,12 +263,12 @@ class MultimediaObjectController extends SortableAdminController implements NewA
           if ($config->isApiRequest()) {
             return $this->handleView($this->view($formMeta));
           }
-          
+
           /*
             $criteria = $this->getCriteria($config);
             $resources = $this->getResources($request, $config, $criteria);
           */
-          
+
           $mms = $this->getListMultimediaObjects($series);
 
           return $this->render('PumukitNewAdminBundle:MultimediaObject:list.html.twig',
@@ -276,7 +276,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
                                      'series' => $series,
                                      'mms' => $mms
                                      )
-                               );         
+                               );
         }
 
         if ($config->isApiRequest()) {
@@ -299,7 +299,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
                                    )
                              );
     }
-    
+
     /**
      * Display the form for editing or update the resource.
      */
@@ -394,7 +394,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
     }
 
     /**
-     * 
+     *
      */
     public function getChildrenTagAction(Tag $tag, Request $request)
     {
@@ -471,7 +471,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
                                        'group' => $n->getPath()
                                        );
         }
-        
+
         $json['prototype'] = $resource->isPrototype();
         return new JsonResponse($json);
     }
@@ -572,17 +572,8 @@ class MultimediaObjectController extends SortableAdminController implements NewA
         $resourceId = $resource->getId();
         $seriesId = $resource->getSeries()->getId();
 
-        if(!$this->isGranted(Permission::MODIFY_OWNER)) {
-            $loggedInUser = $this->getUser();
-            $personService = $this->get('pumukitschema.person');
-            $person = $personService->getPersonFromLoggedInUser($loggedInUser);
-            $role = $personService->getPersonalScopeRole();
-            if( !$person ||
-                !$resource->containsPersonWithRole($person, $role) ||
-                count($resource->getPeopleByRole($role, true)) > 1) {
-                return new Response('You don\'t have enough permissions to delete this mmobj. Contact your administrator.', Response::HTTP_FORBIDDEN);
-            }
-        }
+        if(!$this->isUserAllowedToDelete($resource))
+            return new Response('You don\'t have enough permissions to delete this mmobj. Contact your administrator.', Response::HTTP_FORBIDDEN);
 
         try {
             $this->get('pumukitschema.factory')->deleteMultimediaObject($resource);
@@ -594,7 +585,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
             $this->get('session')->remove('admin/mms/id');
         }
 
-        return $this->redirect($this->generateUrl('pumukitnewadmin_mms_list', 
+        return $this->redirect($this->generateUrl('pumukitnewadmin_mms_list',
                                                   array('seriesId' => $seriesId)));
     }
 
@@ -609,6 +600,8 @@ class MultimediaObjectController extends SortableAdminController implements NewA
         $factory = $this->get('pumukitschema.factory');
         foreach ($ids as $id) {
             $resource = $this->find($id);
+            if(!$this->isUserAllowedToDelete($resource))
+                continue;
             try{
                 $factory->deleteMultimediaObject($resource);
             } catch (\Exception $e) {
@@ -645,7 +638,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
 
         $this->get('pumukitschema.factory')->cloneMultimediaObject($resource);
 
-        return $this->redirect($this->generateUrl('pumukitnewadmin_mms_list', 
+        return $this->redirect($this->generateUrl('pumukitnewadmin_mms_list',
                                                   array('seriesId' => $seriesId)));
     }
 
@@ -697,7 +690,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
                 $update_session = false;
             }
         }
- 
+
         if($update_session){
             $this->get('session')->remove('admin/mms/id');
         }
@@ -785,8 +778,8 @@ class MultimediaObjectController extends SortableAdminController implements NewA
               ->getQuery()
               ->execute();
         }
-        
-        return $this->redirect($this->generateUrl('pumukitnewadmin_mms_list'));      
+
+        return $this->redirect($this->generateUrl('pumukitnewadmin_mms_list'));
     }
 
     /**
@@ -805,7 +798,7 @@ class MultimediaObjectController extends SortableAdminController implements NewA
         $mms = $multimediaObjectRepo->findBySeries($multimediaObject->getSeries())->toArray();
         $tags = $multimediaObject->getTags()->toArray();
         $this->get('pumukitschema.tag')->resetTags($mms, $tags);
-        
+
         return new JsonResponse("");
     }
 
@@ -823,5 +816,27 @@ class MultimediaObjectController extends SortableAdminController implements NewA
         $activeEditor = array_key_exists('pumukit_videoeditor_index', $routes);
 
         return $activeEditor;
+    }
+
+    /**
+     * Returns true if the user has enough permissions to delete the $resource passed
+     *
+     * This function will always return true if the user the MODIFY_ONWER permission. Otherwise,
+     * it checks if it is the owner of the object (and there are no other owners) and returns false if not.
+     */
+    protected function isUserAllowedToDelete(MultimediaObject $resource)
+    {
+        if(!$this->isGranted(Permission::MODIFY_OWNER)) {
+            $loggedInUser = $this->getUser();
+            $personService = $this->get('pumukitschema.person');
+            $person = $personService->getPersonFromLoggedInUser($loggedInUser);
+            $role = $personService->getPersonalScopeRole();
+            if( !$person ||
+                !$resource->containsPersonWithRole($person, $role) ||
+                count($resource->getPeopleByRole($role, true)) > 1) {
+                return false;
+            }
+        }
+        return true;
     }
 }
